Added a barebones basic OOBE
Basically it will detect if the system has an user pre-configured. If the main file ran without a pre-config'ed user, it'll run the OOBE.
Additionally this is also controlled by the "SETUP_IN_PROGRESS" key in SYS_CONFIG. that value should be 0 for an installed shell.

DWM now also attaches the close and minimize buttons to the DWMFrame object which is being returned to the app, to give it control over the buttons such as disabling.