Built foundation for subprocesses, starting with window controls [Read more[
Up untill now, if a host app has a dialogue box, and you minimize the host and restore it, the dialogue box will be gone, forcing you to use task manager's focus option everytime. This changes now, as all dialogue boxes (WHICH ARE ATTACHED BY DWM) Will be focused in/out respectively, and closing
1) the subprocess, will deregister itself from the host pap and then close
2) the host, will close all subprocesses under it

With that, messageboxes also have associate subprocess option, which is yet to be implemented, and it'll be done in the next commit.