[NEW!] Shelve Editor Program!  + DWM Subprocess fixes, change configuration directory to respective Users!
Shelve Editor: It Allows you to edit shelves, removing and adding keys, as well as sub-dictionary views for better nested-dictionary editing.
TO ADD: Creating a NEW shelve.

Fixed subprocesses not quitting if they used the .quit() function, which resulted in the parent process being unaware of the process quit. When the parent is terminated, the now non-existent process is called for quitting, which errors out.  This is fixed.

Moved User configurations from ProgramFiles/<user> to Users/<user> to de-clog the ProgramFiles directory, and only use it for python files and SYS_CONFIGs.