Inter-app communication! [Read below]
Basically this is a system where apps can interact with each other, by basically "messaging" another app.

So for a hypothetical example, say if a process wanted all the running apps from taskmanager, it can send a message to task manager, and if taskmanager is listening, it can reply to the sender the information it asked for.

Currently this is not IMPLEMENTED anywhere but later we can use it in a lot of places.